api_ttm/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-multipart==0.0.20
requests==2.32.3 